+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - The badges JoinDistanceBadge and LeaveDistanceBadge for displaying the leave and join thresholds when viewing an event.
 - A utility function returning a distance formatted to meters or kilometers.
 - Individual leave and join thresholds for events, selectable on the CreateEventForm.
+- A location title to events.
 
 ### Changed